src/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==2.1
django-common-helpers==0.9.2
django-cors-headers==2.5.1
django-cron==0.5.1
django-filter==2.1.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.1
Flask==1.0.2
gitdb2==2.0.6
GitPython==3.0.4
icalendar==4.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
oauthlib==3.0.1
Pillow==6.0.0
ply==3.11
psycopg2==2.7.7
Pygments==2.3.1
PyJWT==1.7.1
pypandoc==1.4
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2018.9
requests==2.22.0
six==1.12.0
smmap2==2.0.5
urllib3==1.24.1
Werkzeug==0.14.1
